Dynamic patrols v0.3
YAY! now we only need to place patrol markers and the game will do the rest (spawn the enemies and give them the patrols).
Enemies can now actually patrol and we can give the the min and max wait time when the enemy arrives at a patrol point.
Good stuff :)